Large commit with several related bug fixes: 1) All fTools functions should now add all layers to dropdowns (including non-visible layers), 2) All fTools functions should now be non-modal 3) Several fTools functions have been spead up slightly 4) Where possible, internal (cpp) functions have be used to replace older Python functions 5) Defining projections now also considers .qpj files Fixes bugs #2502 and #2506, and possibly others?
git-svn-id: http://svn.osgeo.org/qgis/trunk@13037 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ python/plugins/fTools/tools/doPointsInPolygon.py
@@ -104,7 +104,7 @@ class Dialog(QDialog, Ui_Dialog):
         index = polyProvider.fieldNameIndex(unicode(inField))
         if index == -1:
             index = polyProvider.fieldCount()
-            field = QgsField(unicode(inField), QVariant.Int, "real", 24, 15, self.tr("point count field"))
+            field = QgsField(unicode(inField), QVariant.Double, "real", 24, 15, self.tr("point count field"))
             fieldList[index] = field
         sRs = polyProvider.crs()
         check = QFile(self.shapefileName)
